+++ app/src/ai/get_relevant_files/remote_search/native.rs
@@ -124,51 +124,6 @@ pub(super) fn send_request(
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use remote_server::proto::{file_context_proto, FileContextProto};
-    use std::path::PathBuf;
-
-    use super::file_contents_from_response;
-
-    #[test]
-    fn file_contents_from_response_keeps_only_whole_text_files() {
-        let response = remote_server::proto::ReadFileContextResponse {
-            file_contexts: vec![
-                FileContextProto {
-                    file_name: "/repo/src/lib.rs".to_string(),
-                    content: Some(file_context_proto::Content::TextContent(
-                        "content".to_string(),
-                    )),
-                    line_range_start: None,
-                    line_range_end: None,
-                    last_modified_epoch_millis: None,
-                    line_count: 1,
-                },
-                FileContextProto {
-                    file_name: "/repo/src/fragment.rs".to_string(),
-                    content: Some(file_context_proto::Content::TextContent(
-                        "fragment".to_string(),
-                    )),
-                    line_range_start: Some(1),
-                    line_range_end: Some(2),
-                    last_modified_epoch_millis: None,
-                    line_count: 1,
-                },
-            ],
-            failed_files: vec![],
-        };
-
-        let file_contents = file_contents_from_response(response);
-
-        assert_eq!(file_contents.len(), 1);
-        assert_eq!(
-            file_contents.get(&PathBuf::from("/repo/src/lib.rs")),
-            Some(&"content".to_string())
-        );
-    }
-}
-
 // The controller owns request lifecycle concerns like cancellation, pending request tracking, and
 // result emission. This function only contains the remote-specific pipeline: store content hashes
 // -> daemon fragment metadata -> remote file reads -> fragment reranking.
@@ -499,3 +454,7 @@ fn remote_availability_failure(
         },
     }
 }
+
+#[cfg(test)]
+#[path = "native_tests.rs"]
+mod tests;
